Move GameSelector to its own package in anticipation of further refactor to break-up the class to smaller pieces.
