+++ game-core/src/main/java/games/strategy/engine/framework/network/ui/ChangeGameToSaveGameClientAction.java
@@ -5,7 +5,7 @@ import java.io.File;
 
 import javax.swing.AbstractAction;
 
-import games.strategy.engine.framework.startup.ui.panels.main.GameSelectorPanel;
+import games.strategy.engine.framework.startup.ui.panels.main.game.selector.GameSelectorPanel;
 import games.strategy.net.IClientMessenger;
 
 /**
